dealmart/requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.4.2
dataclasses==0.6
Django==2.1.4
django-filter==2.0.0
django-mathfilters==0.4.0
django-otp==0.5.1
django-phonenumber-field==2.0.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djongo==1.2.30
ecdsa==0.13
future==0.17.1
httpie==1.0.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
phonenumbers==8.9.14
Pillow==5.2.0
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.7.2
python-jose==3.0.1
pytz==2018.5
PyYAML==3.13
requests==2.20.1
rsa==4.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.0.0
